refactor: split completer.rs into string_context, meta, and r_completer modules
Extract the 2372-line completer.rs into focused modules:
- string_context.rs: tree-sitter string detection + path-in-string completion
- meta.rs: MetaCommandCompleter and all meta command logic
- r_completer.rs: RCompleter with namespace fuzzy completion
- completer.rs: slimmed to CombinedCompleter + re-exports

No behavioral changes; all tests moved alongside their code.
